Refactor tailwind suggestions
