+++ apps/studio/src/routes/editor/EditPanel/StylesTab/single/TailwindInput/index.tsx
@@ -1,0 +1,193 @@
+import { useEditorEngine } from '@/components/Context';
+import { sendAnalytics } from '@/lib/utils';
+import type { CodeDiffRequest } from '@onlook/models/code';
+import { MainChannels } from '@onlook/models/constants';
+import type { TemplateNode } from '@onlook/models/element';
+import { Icons } from '@onlook/ui/icons';
+import { Textarea } from '@onlook/ui/textarea';
+import { observer } from 'mobx-react-lite';
+import { useEffect, useRef, useState } from 'react';
+import { SuggestionsList, type SuggestionsListRef } from './AutoComplete';
+
+const TailwindInput = observer(() => {
+    const editorEngine = useEditorEngine();
+
+    const instanceRef = useRef<HTMLTextAreaElement>(null);
+    const [instance, setInstance] = useState<TemplateNode | undefined>();
+    const [instanceClasses, setInstanceClasses] = useState<string>('');
+    const [isInstanceFocused, setIsInstanceFocused] = useState(false);
+
+    const rootRef = useRef<HTMLTextAreaElement>(null);
+    const [root, setRoot] = useState<TemplateNode | undefined>();
+    const [rootClasses, setRootClasses] = useState<string>('');
+    const [isRootFocused, setIsRootFocused] = useState(false);
+
+    const suggestionRef = useRef<SuggestionsListRef>(null);
+    const [showSuggestions, setShowSuggestions] = useState(true);
+
+    const handleInput = (
+        e: React.FormEvent<HTMLTextAreaElement>,
+        setClasses: React.Dispatch<React.SetStateAction<string>>,
+    ) => {
+        const value = e.currentTarget.value;
+        setClasses(value);
+        suggestionRef.current?.handleInput(value);
+    };
+
+    const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+        if (showSuggestions) {
+            suggestionRef.current?.handleKeyDown(e);
+        } else if (e.key === 'Enter' || e.key === 'Tab' || e.key === 'Escape') {
+            e.currentTarget.blur();
+            e.preventDefault();
+        }
+    };
+
+    useEffect(() => {
+        if (editorEngine.elements.selected.length) {
+            const selectedEl = editorEngine.elements.selected[0];
+            getInstanceClasses(selectedEl.selector);
+            getRootClasses(selectedEl.selector);
+        }
+    }, [editorEngine.elements.selected]);
+
+    async function getInstanceClasses(selector: string) {
+        const instance = editorEngine.ast.getInstance(selector);
+        setInstance(instance);
+        if (instance) {
+            const instanceClasses: string[] = await window.api.invoke(
+                MainChannels.GET_TEMPLATE_NODE_CLASS,
+                instance,
+            );
+            setInstanceClasses(instanceClasses.join(' '));
+        }
+    }
+
+    async function getRootClasses(selector: string) {
+        const root = editorEngine.ast.getRoot(selector);
+        setRoot(root);
+        if (root) {
+            const rootClasses: string[] = await window.api.invoke(
+                MainChannels.GET_TEMPLATE_NODE_CLASS,
+                root,
+            );
+            setRootClasses(rootClasses.join(' '));
+        }
+    }
+
+    const createCodeDiffRequest = async (templateNode: TemplateNode, className: string) => {
+        const request: CodeDiffRequest = {
+            templateNode,
+            selector: editorEngine.elements.selected[0].selector,
+            attributes: { className },
+            insertedElements: [],
+            movedElements: [],
+            removedElements: [],
+            groupElements: [],
+            ungroupElements: [],
+            overrideClasses: true,
+        };
+        const res = await editorEngine.code.getAndWriteCodeDiff([request]);
+        if (res) {
+            sendAnalytics('tailwind action');
+        }
+    };
+
+    const adjustHeight = (textarea: HTMLTextAreaElement) => {
+        textarea.style.height = 'auto';
+        textarea.style.height = `${textarea.scrollHeight + 20}px`;
+    };
+
+    useEffect(() => {
+        if (instanceRef.current) {
+            adjustHeight(instanceRef.current);
+        }
+    }, [instanceClasses]);
+
+    useEffect(() => {
+        if (rootRef.current) {
+            adjustHeight(rootRef.current);
+        }
+    }, [rootClasses]);
+
+    const EnterIndicator = () => {
+        return (
+            <div className="absolute bottom-1 right-2 text-xs text-gray-500 flex items-center">
+                <span>enter to apply</span>
+                <Icons.Reset className="ml-1" />
+            </div>
+        );
+    };
+
+    return (
+        <div className="flex flex-col gap-2 text-xs text-foreground-onlook">
+            {instance && <p>Instance</p>}
+            {instance && (
+                <div className="relative">
+                    <div>
+                        <Textarea
+                            ref={instanceRef}
+                            className="w-full text-xs text-foreground-active break-normal bg-background-onlook/75 focus-visible:ring-0"
+                            placeholder="Add tailwind classes here"
+                            value={instanceClasses}
+                            onInput={(e) => handleInput(e, setInstanceClasses)}
+                            onKeyDown={handleKeyDown}
+                            onBlur={(e) => {
+                                setShowSuggestions(false);
+                                setIsInstanceFocused(false);
+                                instance && createCodeDiffRequest(instance, e.target.value);
+                            }}
+                            onFocus={() => setIsInstanceFocused(true)}
+                        />
+                        {isInstanceFocused && (
+                            <SuggestionsList
+                                currentInput={instanceClasses}
+                                showSuggestions={showSuggestions}
+                                setCurrentInput={setInstanceClasses}
+                                ref={suggestionRef}
+                                setShowSuggestions={setShowSuggestions}
+                                setClasses={setInstanceClasses}
+                            />
+                        )}
+                    </div>
+                    {isInstanceFocused && <EnterIndicator />}
+                </div>
+            )}
+
+            {instance && root && <p>Component</p>}
+            {root && (
+                <div className="relative">
+                    <div>
+                        <Textarea
+                            ref={rootRef}
+                            className="w-full text-xs text-foreground-active break-normal bg-background-onlook/75 focus-visible:ring-0 resize-none"
+                            placeholder="Add tailwind classes here"
+                            value={rootClasses}
+                            onInput={(e) => handleInput(e, setRootClasses)}
+                            onKeyDown={handleKeyDown}
+                            onBlur={(e) => {
+                                setShowSuggestions(false);
+                                setIsRootFocused(false);
+                                root && createCodeDiffRequest(root, e.target.value);
+                            }}
+                            onFocus={() => setIsRootFocused(true)}
+                        />
+                        {isRootFocused && (
+                            <SuggestionsList
+                                ref={suggestionRef}
+                                showSuggestions={showSuggestions}
+                                currentInput={rootClasses}
+                                setCurrentInput={setRootClasses}
+                                setShowSuggestions={setShowSuggestions}
+                                setClasses={setRootClasses}
+                            />
+                        )}
+                    </div>
+                    {isRootFocused && <EnterIndicator />}
+                </div>
+            )}
+        </div>
+    );
+});
+
+export default TailwindInput;
